+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Remove old xml code from 2014 no longer user
 - Make launcher compatable with Java 9 and newer
 - Change ram settings to be spinners
+- Warn when using Java 9 or newer that some packs may not work